I was mistaken about how ggml shards the model.
It splits each layer across all files. This has the massive disadvantage that it
makes it impossible to load layers from the file one at a time. I'm sure there's
some reason for it though.

In the meantime let's get the one file version working and then I'll deal with
sharding.